requirements.txt
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio @ file:///croot/anyio_1729121277521/work
attrs @ file:///croot/attrs_1734533101012/work
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
datasets @ file:///croot/datasets_1716911606380/work
dill @ file:///croot/dill_1715094664823/work
distro @ file:///croot/distro_1714488253808/work
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
frozenlist @ file:///croot/frozenlist_1730902802621/work
fsspec @ file:///croot/fsspec_1714461537038/work
h11 @ file:///croot/h11_1706652277403/work
httpcore @ file:///croot/httpcore_1706728464539/work
httpx @ file:///croot/httpx_1723474802858/work
huggingface_hub @ file:///croot/huggingface_hub_1724853938404/work
idna @ file:///croot/idna_1714398848350/work
jiter @ file:///croot/jiter_1729808929109/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
multidict @ file:///croot/multidict_1730905498140/work
multiprocess @ file:///work/perseverance-python-buildout/croot/multiprocess_1698870874259/work
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
openai @ file:///croot/openai_1738233361885/work
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp312-cp312-linux_x86_64.whl#sha256=57b66702d418720ec8483f7c4ec7c08d41815316ad7ce09d5b7bbc34eefcfdfd
propcache @ file:///croot/propcache_1732303986938/work
pyarrow @ file:///croot/pyarrow_1738229418212/work/python
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1730999120400/work
safetensors @ file:///croot/safetensors_1732227615904/work
setuptools==75.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///croot/sniffio_1705431295498/work
tokenizers @ file:///croot/tokenizers_1732129617059/work
tqdm @ file:///croot/tqdm_1738943501192/work
transformers @ file:///croot/transformers_1729767158381/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1737133630106/work
wheel==0.45.1
xxhash @ file:///croot/python-xxhash_1737039434400/work
yarl @ file:///croot/yarl_1732546845924/work
